Always use latest version of Boundary
This ensures that the latest version of Boundary is always used when
it's explicitly specified in code (some examples, like AWS, currently
rely on a local binary to upload, so those were not altered).

It also adds some explicit CORS rules for the Docker example, where it's
necessary to ensure that the Boundary UI now works.

Finally, some notes in the Docker example have been added to note that
you may need an entry in your /etc/hosts file to ensure that you can
connect to the workers.